requirements.txt
mysql-replication
kafka-python
redis
pyelasticsearch
pytz
